pyproject.toml
[tool.poetry]
name = "log-api"
version = "0.1.0"
description = "Technical challenge for selective process."
authors = ["Mateus Oliveira"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.78.0"
SQLAlchemy = "^1.4.39"
uvicorn = "^0.18.2"
psycopg2 = "^2.9.3"
python-decouple = "^3.6"
email-validator = "^1.2.1"
PyJWT = "^2.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
bandit = "^1.7.4"
black = "^22.3.0"
editorconfig-checker = "^2.4.0"
isort = "^5.10.1"
mypy = "^0.961"
pre-commit = "^2.19.0"
prospector = "^1.7.7"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
safety = "^1.10.3"
Sphinx = "^5.0.2"
sphinx-rtd-theme = "^1.0.0"
toml = "^0.10.2"
types-toml = "^0.10.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = "error"
addopts = [
    "-vv",
    "-s",
    "--cov-report=term-missing",
    "--cov-report=html:tests/coverage-results/htmlcov",
    "--cov=source",
    # TODO finish testing
    # "--cov-fail-under=100",
]

[tool.mypy]
python_version = "3.7"
strict = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "decouple",
    "sqlalchemy.*",
    "uvicorn.*",
    "passlib.*",
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
line_length = 79
verbose = true
multi_line_output = 3
include_trailing_comma = true
skip_gitignore = true

[tool.black]
line-length = 79
verbose = true
